V4: security + data-consistency hardening, Alembic, bulk ops & UX fixes
在 V3(混合路由 / 序列 KG 佇列)基礎上,依深度稽核逐項修復並強化:

安全
- 修補註冊提權(role 白名單)、source_pdf_path 檔案外洩(LFI)
- chunk / 資料夾寫入改 admin-only
- refresh token 原子輪替 + 重用偵測
- rate limit:proxy-aware IP key + 帳號層退避 + 重運算端點配額

資料一致性
- 刪除 / 重跑 OCR / 向量化 改原子順序(先建後換,避免資料遺失)
- partial ingest 修正(先 embed 成功才寫 chunks)
- FAISS 原子寫(.tmp+os.replace)+ 從 DB 重建 index 端點
- embedding 截斷改設定化(EMBEDDING_MAX_CHARS)、超長 segment 強切
- 刪文件同步清 KG 邊 / 結構節點 / user analysis / notes

架構 / 穩定性
- 導入 Alembic 取代啟動時 ad-hoc DDL;faiss_id 改 BigInteger(PostgreSQL 相容)
- 重啟後 stale 背景任務復原;KG worker 加固;SQLite WAL + busy_timeout
- 同文件重跑去重(避免重複觸發並行衝突)

前端
- 全域 ErrorBoundary、換帳號身分修正、SSE 斷線 fallback、任務輪詢完成即停 + 多分頁同步
- 路由 code-splitting;antd 棄用屬性、zustand named import
- 參考來源卡片修復(同步 ref)+ 來源片段 Markdown/表格渲染
- 文件列表多選批次(分類 / 移到資料夾 / 刪除)
- PDF 分析對話自動捲到最新一次查詢

驗證:scripts/verify_audit_fixes.py(23 項單元)+ 端對端測試全通過;前端 lint/build 通過。

+++ backend/app/models.py
@@ -1,7 +1,7 @@
 import uuid
 from typing import List, Optional
 
-from sqlalchemy import Boolean, DateTime, Float, ForeignKey, Index, Integer, JSON, String, Text, UniqueConstraint
+from sqlalchemy import BigInteger, Boolean, DateTime, Float, ForeignKey, Index, Integer, JSON, String, Text, UniqueConstraint
 from sqlalchemy.ext.mutable import MutableDict, MutableList
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from sqlalchemy.sql import func
@@ -219,7 +219,9 @@ class DocumentChunk(BaseMixin, Base):
     paragraph_index: Mapped[Optional[int]] = mapped_column(Integer)
     text: Mapped[str] = mapped_column(Text, nullable=False)
     embedding: Mapped[List[float]] = mapped_column(MutableList.as_mutable(JSON), default=list, nullable=False)
-    faiss_id: Mapped[int] = mapped_column(Integer, unique=True, nullable=False)
+    # BigInteger: faiss_id 是 uuid4 % 2^63，SQLite 沒事但 PostgreSQL 的 INTEGER 是 32-bit
+    # 會 NumericValueOutOfRange（audit H6）。BigInteger 在 SQLite 仍映射為 INTEGER，相容。
+    faiss_id: Mapped[int] = mapped_column(BigInteger, unique=True, nullable=False)
 
     document: Mapped[Document] = relationship("Document", back_populates="chunks")
